+++ src/components/Navigation/LoggedIn.js
@@ -4,13 +4,15 @@ import { logOut } from "../../store/user/actions";
 import Button from "react-bootstrap/Button";
 import { selectUser } from "../../store/user/selectors";
 import Nav from "react-bootstrap/Nav";
+import NavbarItem from "./NavbarItem";
 
 export default function LoggedIn() {
   const dispatch = useDispatch();
   const user = useSelector(selectUser);
   return (
     <>
-      <Nav.Item style={{ padding: ".5rem 1rem" }}>{user.email}</Nav.Item>
+      <NavbarItem path="/import" linkText="Import Photos" />
+      <Nav.Item style={{ padding: ".5rem 1rem" }}>{user.name}</Nav.Item>
       <Button onClick={() => dispatch(logOut())}>Logout</Button>
     </>
   );
