+++ src/components/Navigation/index.js
@@ -22,7 +22,7 @@ export default function Navigation() {
       <Navbar.Collapse id="basic-navbar-nav">
         <Nav style={{ width: "100%" }} fill>
           <NavbarItem path="/" linkText="Home" />
-          <NavbarItem path="/other" linkText="Other" />
+          {/* <NavbarItem path="/import" linkText="Import photos" /> */}
           {loginLogoutControls}
         </Nav>
       </Navbar.Collapse>
